tests/wpt/tests/html/semantics/forms/the-select-element/stylable-select/select-keyboard-hidden-option.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/whatwg/html/issues/9799">
<link rel=help href="https://issues.chromium.org/issues/360357707">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<style>
select, select::picker(select) {
  appearance: base-select;
}
</style>

<select>
  <option hidden>hidden one</option>
  <option class=two>two</option>
  <option class=three>three</option>
</select>

<script>
const ArrowDown = '\uE015';

promise_test(async () => {
  const select = document.querySelector('select');
  const hiddenOption = document.querySelector('option[hidden]');
  const optionTwo = document.querySelector('option.two');
  const optionThree = document.querySelector('option.three');

  select.focus();
  await test_driver.send_keys(select, ' ');
  assert_true(select.matches(':open'),
    'Select should open after pressing spacebar.');

  assert_equals(document.activeElement, optionTwo,
    'The first visible <option> should get focused after opening picker.');

  await (new test_driver.Actions()
    .keyDown(ArrowDown)
    .keyUp(ArrowDown))
    .send();

  assert_equals(document.activeElement, optionThree,
    'The second visible <option> should get focused after arrowdown.');
}, 'Hidden <option>s should not get initial focus.');
</script>
